zappa/requirements.txt
zappa==0.51.0
fastapi==0.52.0
aiohttp==3.6.2
Pillow==7.0.0
uvicorn==0.11.3
mangum==0.7.4